+++ tests/models/test_tag.py
@@ -72,6 +72,7 @@ class TagTest(ModelTest):
                     'name': tag_name,
                     'subscriptions': [
                         {
+                            'tags': ['test'],
                             'contacts': ['3c01399e-1d40-46dd-934f-318e8255fd3e'],
                             'enabled': True,
                             'id': subscription_id,
@@ -86,7 +87,7 @@ class TagTest(ModelTest):
                             ],
                             'endOffset': 1439,
                             'startOffset': 0,
-                            'tzOffset': -180
+                            'tzOffset': 0
                         }
                     ],
                     'triggers': []
@@ -111,12 +112,13 @@ class TagTest(ModelTest):
         tag_name = 'tag_name'
         subscription_id = '3c01399e-1d40-46dd-934f-318e8255fd3e'
         trigger_id = '123'
-        stats = {
+        return_value = {
             'list': [
                 {
                     'name': tag_name,
                     'subscriptions': [
                         {
+                            'tags': ['test'],
                             'contacts': ['1'],
                             'enabled': True,
                             'id': subscription_id,
@@ -131,7 +133,7 @@ class TagTest(ModelTest):
                             ],
                             'endOffset': 1439,
                             'startOffset': 0,
-                            'tzOffset': -180
+                            'tzOffset': 0
                         }
                     ],
                     'triggers': [trigger_id]
@@ -139,26 +141,12 @@ class TagTest(ModelTest):
             ]
         }
 
-        state = {
-            'state': 'OK',
-            'trigger_id': trigger_id
-            }
-
-        trigger = {
-            'id': trigger_id,
-            'name': 'name',
-            'tags': ['tag'],
-            'targets': ['pattern'],
-            'warn_value': 0,
-            'error_value': 1
-            }
-
-        with patch.object(client, 'get', side_effect=[stats, state, trigger]) as get_mock:
-            triggers = tag_manager.fetch_assigned_triggers(tag_name)
+        with patch.object(client, 'get', return_value=return_value) as get_mock:
+            triggerIds = tag_manager.fetch_assigned_triggers(tag_name)
 
             self.assertTrue(get_mock.called)
-            get_mock.assert_called_with('trigger/123')
-            self.assertEqual(1, len(triggers))
+            get_mock.assert_called_with('tag/stats')
+            self.assertEqual(1, len(triggerIds))
 
     def test_fetch_assigned_subscriptions(self):
         client = Client(self.api_url)
@@ -173,6 +161,7 @@ class TagTest(ModelTest):
                     'name': tag_name,
                     'subscriptions': [
                         {
+                            'tags': ['test'],
                             'contacts': ['1'],
                             'enabled': True,
                             'id': subscription_id,
@@ -187,7 +176,7 @@ class TagTest(ModelTest):
                             ],
                             'endOffset': 1439,
                             'startOffset': 0,
-                            'tzOffset': -180
+                            'tzOffset': 0
                         }
                     ],
                     'triggers': [trigger_id]