+++ tests/test_client.py
@@ -9,7 +9,11 @@ from moira_client.client import Client
 from moira_client.client import InvalidJSONError
 
 TEST_API_URL = 'http://test/api/url'
-AUTH_HEADERS = {'X-Webauth-User': 'login'}
+TEST_HEADERS = {
+    'X-Webauth-User': 'login',
+    'Content-Type': 'application/json',
+    'User-Agent': 'Python Moira Client'
+    }
 
 
 class FakeResponse:
@@ -35,12 +39,12 @@ class ClientTest(unittest.TestCase):
         with patch.object(requests, 'get', side_effects=get) as mock_get:
             test_path = 'test_path'
 
-            client = Client(TEST_API_URL, AUTH_HEADERS)
+            client = Client(TEST_API_URL, TEST_HEADERS)
             client.get(test_path)
 
         self.assertTrue(mock_get.called)
         expected_url_call = TEST_API_URL + '/' + test_path
-        mock_get.assert_called_with(expected_url_call, headers=AUTH_HEADERS, auth=None)
+        mock_get.assert_called_with(expected_url_call, headers=TEST_HEADERS, auth=None)
 
     def test_put(self):
 
@@ -51,12 +55,12 @@ class ClientTest(unittest.TestCase):
             test_path = 'test_path'
             test_data = {'test': 'test'}
 
-            client = Client(TEST_API_URL, AUTH_HEADERS)
+            client = Client(TEST_API_URL, TEST_HEADERS)
             client.put(test_path, data=test_data)
 
         self.assertTrue(mock_put.called)
         expected_url_call = TEST_API_URL + '/' + test_path
-        mock_put.assert_called_with(expected_url_call, data=test_data, headers=AUTH_HEADERS, auth=None)
+        mock_put.assert_called_with(expected_url_call, data=test_data, headers=TEST_HEADERS, auth=None)
 
     def test_delete(self):
 
@@ -66,12 +70,12 @@ class ClientTest(unittest.TestCase):
         with patch.object(requests, 'delete', side_effects=delete) as mock_delete:
             test_path = 'test_path'
 
-            client = Client(TEST_API_URL, AUTH_HEADERS)
+            client = Client(TEST_API_URL, TEST_HEADERS)
             client.delete(test_path)
 
         self.assertTrue(mock_delete.called)
         expected_url_call = TEST_API_URL + '/' + test_path
-        mock_delete.assert_called_with(expected_url_call, headers=AUTH_HEADERS, auth=None)
+        mock_delete.assert_called_with(expected_url_call, headers=TEST_HEADERS, auth=None)
 
     def test_get_invalid_response(self):
 
@@ -83,13 +87,13 @@ class ClientTest(unittest.TestCase):
         with patch.object(requests, 'get', side_effects=get, return_value=response) as mock_get:
             test_path = 'test_path'
 
-            client = Client(TEST_API_URL, AUTH_HEADERS)
+            client = Client(TEST_API_URL, TEST_HEADERS)
             with self.assertRaises(InvalidJSONError):
                 client.get(test_path)
 
         self.assertTrue(mock_get.called)
         expected_url_call = TEST_API_URL + '/' + test_path
-        mock_get.assert_called_with(expected_url_call, headers=AUTH_HEADERS, auth=None)
+        mock_get.assert_called_with(expected_url_call, headers=TEST_HEADERS, auth=None)
 
     def test_put_invalid_response(self):
         test_data = {'test': 'test'}
@@ -102,13 +106,13 @@ class ClientTest(unittest.TestCase):
         with patch.object(requests, 'put', side_effects=put, return_value=response) as mock_put:
             test_path = 'test_path'
 
-            client = Client(TEST_API_URL, AUTH_HEADERS)
+            client = Client(TEST_API_URL, TEST_HEADERS)
             with self.assertRaises(InvalidJSONError):
                 client.put(test_path, data=test_data)
 
         self.assertTrue(mock_put.called)
         expected_url_call = TEST_API_URL + '/' + test_path
-        mock_put.assert_called_with(expected_url_call, data=test_data, headers=AUTH_HEADERS, auth=None)
+        mock_put.assert_called_with(expected_url_call, data=test_data, headers=TEST_HEADERS, auth=None)
 
     def test_delete_invalid_response(self):
 
@@ -120,10 +124,10 @@ class ClientTest(unittest.TestCase):
         with patch.object(requests, 'delete', side_effects=delete, return_value=response) as mock_delete:
             test_path = 'test_path'
 
-            client = Client(TEST_API_URL, AUTH_HEADERS)
+            client = Client(TEST_API_URL, TEST_HEADERS)
             with self.assertRaises(InvalidJSONError):
                 client.delete(test_path)
 
         self.assertTrue(mock_delete.called)
         expected_url_call = TEST_API_URL + '/' + test_path
-        mock_delete.assert_called_with(expected_url_call, headers=AUTH_HEADERS, auth=None)
+        mock_delete.assert_called_with(expected_url_call, headers=TEST_HEADERS, auth=None)