+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 2.4
+
+- Python Moira client is now MIT licensed
+- Changed default timezone offset to UTC
+- Added dictionary attribute with plotting settings to subscription and<br/>
+  corresponding methods enable_plotting(theme) and disable_plotting()
+- Added boolean attribute mute_new_metrics to trigger
+- Optimized tag stats methods (stats, fetch_assigned_triggers, fetch_assigned_triggers_by_tags)<br/>
+to return trigger id's instead of triggers to speed up method execution time
+
 # 2.3.5
 - Fixed url to update existing subscription